Relocate event subscriber into Listeners, following laravel's convention.
