+++ src/LaravelDDDServiceProvider.php
@@ -3,7 +3,7 @@
 namespace Lunarstorm\LaravelDDD;
 
 use Illuminate\Support\Facades\Event;
-use Lunarstorm\LaravelDDD\Support\CacheClearSubscriber;
+use Lunarstorm\LaravelDDD\Listeners\CacheClearSubscriber;
 use Lunarstorm\LaravelDDD\Support\DomainAutoloader;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
