+++ src/Listeners/CacheClearSubscriber.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Lunarstorm\LaravelDDD\Support;
+namespace Lunarstorm\LaravelDDD\Listeners;
 
 use ErrorException;
 use Illuminate\Events\Dispatcher;
@@ -13,13 +13,13 @@ class CacheClearSubscriber
 
     public function handle(): void
     {
-        $files = glob(base_path(config('ddd.cache_directory').'/ddd-*.php'));
+        $files = glob(base_path(config('ddd.cache_directory') . '/ddd-*.php'));
 
         foreach ($files as $file) {
             try {
                 unlink($file);
             } catch (ErrorException $exception) {
-                if (! str_contains($exception->getMessage(), 'No such file or directory')) {
+                if (!str_contains($exception->getMessage(), 'No such file or directory')) {
                     dump($exception->getMessage());
                     throw $exception;
                 }
